typing corrections for AnimatedImage coordinates

+++ include/sampleClasses/Apple.h
@@ -5,7 +5,7 @@ class Apple : public AnimatedImage{
     private:
         Collider collider;
     public:
-        Apple(std::string imgId,int x,int y,std::vector<std::vector<int>> imageCoords,float imageCoordsIndex,float animationSpeed,std::vector<std::string> idsList,std::vector<std::string> scriptsIds):
+        Apple(std::string imgId,int x,int y,std::vector<std::vector<float>> imageCoords,float imageCoordsIndex,float animationSpeed,std::vector<std::string> idsList,std::vector<std::string> scriptsIds):
             AnimatedImage(imgId,x,y,imageCoords,imageCoordsIndex,animationSpeed,idsList,scriptsIds),
             collider(x,y,width,height)
         {}
